+++ main.py
@@ -1,12 +1,20 @@
 from dane import users_list
 
-from UTILS.my_functions import add_user_to, remove_user_from, show_users_from, gui
 
 from UTILS.my_functions import gui
 
 
 gui(users_list)
-
+# TO DO wlaczyc do kodu aktualnego obsluge baz danych
+# TO DO dodac tabele do bazy danych reprezentujaca uzytkownika
+# TO DO napisac klase uzytkownika o strukturze zgodnej z ta zdefiniowaną w my_data
+# TODO oddac sprawozdanie do dnia 20.12.2023 (sroda)
+# TODO MA ZAWIERAC str tytulowa, spis tresci, opis kodu realizojuacego zadanie rysowania mapy uwzgledniajacy poszczegolne funkcje i ich skladnie
+# TODO zrealizowany kod na git (link) , podsumowanie wnioski koncowe scy i zdjecia mile widziane
+ # class User:
+    # def __init__(self,name,posts,miejscowosc):
+        #self.nazwa = name
+    #def add_user(self):
